requirements.txt
discord.py>=2.3.0
python-dotenv
colorama
ffmpeg-python
requests
loguru
PyNaCl
aiosqlite
Pillow